Overhaul the global simulation object map (again) and start building out event stream and world map functionality

+++ market/data/resources.py
@@ -1,16 +1,18 @@
 import copy
 from random import randint, uniform
 
-from market.data.core import SimulatedObject, loadable
+from market.core import SimulatedObject, loadable
 
 
 @loadable
 class Resource(SimulatedObject):
+
     DEFAULT_DROP_RATE = 50
 
     def __init__(self, name, drop_table=None):
         super(Resource, self).__init__(name)
-        self._drop_table = drop_table or {}
+        if not self.initialized:
+            self._drop_table = drop_table or {}
 
     def generate_drops(self):
         drops = {}
@@ -59,7 +61,7 @@ class Resource(SimulatedObject):
             Equality comparisons for drops should check the represented item
             and both upper and lower bounds for drop amounts.
 
-            Item objects, like all SimulatedObjects, are compared by UUID.
+            Item objects, like all SimulatedObjects, are compared by name.
 
             :param o: the drop to compare against
             :type o: Resource.Drop
@@ -71,21 +73,6 @@ class Resource(SimulatedObject):
         class BoundingError(Exception):
             pass
 
-    class InvalidResourceType(Exception):
-        pass
 
-
-@loadable
-class Zone(SimulatedObject):
-    def __init__(self, name, resources=None):
-        super(Zone, self).__init__(name)
-        self._resources = resources or {}
-
-    def add_resource(self, resource, amount=1):
-        self._resources[resource] = amount
-
-    def remove_resource(self, resource):
-        self._resources.pop(resource)
-
-    def get_resources(self):
-        return copy.deepcopy(self._resources)
+class InvalidResourceType(Exception):
+    pass
